+++ tests/testthat/test_AIC.R
@@ -8,6 +8,6 @@ test_that("The AIC is reproducible", {
   expect_error(AIC(fits["SFO", ]), "column object")
   expect_error(BIC(fits["SFO", ]), "column object")
   expect_equivalent(BIC(fits[, "FOCUS_C"]),
-                    data.frame(df = c(3, 4, 5, 5), AIC = c(59.9, 45.5, 30.0, 40.2)),
+                    data.frame(df = c(3, 4, 5, 5), BIC = c(59.9, 45.5, 30.0, 40.2)),
                     scale = 1, tolerance = 0.1)
 })